Remove Vulkan structure enum to integer casts
VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROVOKING_VERTEX_FEATURES_EXT is now
defined in vulkan_core.h so we no longer have to cast the
VkStructureType to an integer type to avoid warnings/errors about
using a case label value that is not part of the enum.

Remaining cases of such a cast are for Android internal structure type
values. The cast to long has been replaced by a cast to int since C++
states that if the base type of an enum is not declared it "is not
larger than int unless the value of an enumerator cannot fit in an int
or unsigned int", and long could be larger.

Bug: angleproject:3430
Reviewed-on: https://swiftshader-review.googlesource.com/c/SwiftShader/+/62508
Kokoro-Result: kokoro <noreply+kokoro@google.com>
Tested-by: Nicolas Capens <nicolascapens@google.com>

+++ src/Device/Context.cpp
@@ -345,10 +345,7 @@ GraphicsState::GraphicsState(const Device *device, const VkGraphicsPipelineCreat
 	const VkBaseInStructure *extensionCreateInfo = reinterpret_cast<const VkBaseInStructure *>(rasterizationState->pNext);
 	while(extensionCreateInfo)
 	{
-		// Casting to a long since some structures, such as
-		// VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROVOKING_VERTEX_FEATURES_EXT
-		// are not enumerated in the official Vulkan header
-		switch((long)(extensionCreateInfo->sType))
+		switch(extensionCreateInfo->sType)
 		{
 		case VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_LINE_STATE_CREATE_INFO_EXT:
 			{
